@@ -10,7 +10,9 @@
 import $ from 'jquery';
 import ko from 'knockout';
 import mutate from 'mutate-cow';
-import * as ReactDOM from 'react-dom';
+// $FlowIgnore[missing-export]
+import {flushSync} from 'react-dom';
+import * as ReactDOMClient from 'react-dom/client';
 import {createStore} from 'redux';
 
 import FormRowSelectList from '../../../components/FormRowSelectList';
@@ -291,34 +293,37 @@ function removeLanguage(i) {
 
 const getSelectField = field => field;
 
+const workLanguagesNode = document.getElementById('work-languages-editor');
+if (!workLanguagesNode) {
+  throw new Error('Mount point #work-languages-editor does not exist');
+}
+const workLanguagesRoot = ReactDOMClient.createRoot(workLanguagesNode);
+
 function renderWorkLanguages() {
-  const workLanguagesNode = document.getElementById('work-languages-editor');
-  if (!workLanguagesNode) {
-    throw new Error('Mount point #work-languages-editor does not exist');
-  }
   const form: WorkForm = store.getState();
   const selectedLanguageIds =
     form.field.languages.field.map(lang => String(lang.value));
-  ReactDOM.render(
-    <FormRowSelectList
-      addId="add-language"
-      addLabel={l('Add Language')}
-      getSelectField={getSelectField}
-      hideAddButton={
-        selectedLanguageIds.includes(String(LANGUAGE_MUL_ID)) ||
-        selectedLanguageIds.includes(String(LANGUAGE_ZXX_ID))
-      }
-      label={l('Lyrics Languages')}
-      onAdd={addLanguage}
-      onEdit={editLanguage}
-      onRemove={removeLanguage}
-      options={workLanguageOptions}
-      removeClassName="remove-language"
-      removeLabel={l('Remove Language')}
-      repeatable={form.field.languages}
-    />,
-    workLanguagesNode,
-  );
+  flushSync(() => {
+    workLanguagesRoot.render(
+      <FormRowSelectList
+        addId="add-language"
+        addLabel={l('Add Language')}
+        getSelectField={getSelectField}
+        hideAddButton={
+          selectedLanguageIds.includes(String(LANGUAGE_MUL_ID)) ||
+          selectedLanguageIds.includes(String(LANGUAGE_ZXX_ID))
+        }
+        label={l('Lyrics Languages')}
+        onAdd={addLanguage}
+        onEdit={editLanguage}
+        onRemove={removeLanguage}
+        options={workLanguageOptions}
+        removeClassName="remove-language"
+        removeLabel={l('Remove Language')}
+        repeatable={form.field.languages}
+      />,
+    );
+  });
 }
 
 store.subscribe(renderWorkLanguages);
