@@ -9,7 +9,9 @@
 
 import mutate from 'mutate-cow';
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+// $FlowIgnore[missing-export]
+import {flushSync} from 'react-dom';
+import * as ReactDOMClient from 'react-dom/client';
 import * as Sentry from '@sentry/browser';
 
 import {SanitizedCatalystContext} from '../context';
@@ -122,12 +124,21 @@ export default function hydrate<
           if (__DEV__) {
             checkForUnsanitizedEditorData((props: any));
           }
-          ReactDOM.hydrate(
-            <SanitizedCatalystContext.Provider value={$c}>
-              <Component $c={$c} {...props} />
-            </SanitizedCatalystContext.Provider>,
-            root,
-          );
+          /*
+           * Flush updates to the DOM immediately to try and avoid hydration
+           * errors due to user scripts modifying the page. This is ultimately
+           * affected by the order in which the scripts run, though.
+           */
+          flushSync(() => {
+            ReactDOMClient.hydrateRoot(
+              root,
+              <React.StrictMode>
+                <SanitizedCatalystContext.Provider value={$c}>
+                  <Component $c={$c} {...props} />
+                </SanitizedCatalystContext.Provider>
+              </React.StrictMode>,
+            );
+          });
         }
       }
     });
